[Silver IV] Title: 동전 0, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/11047. 동전 0/동전 0.cc
@@ -12,7 +12,7 @@ int main()
 
     std::cin >> n >> k;
 
-    std::vector<long> values(n);
+    std::vector<int> values(n);
 
     for (int i = 0; i < n; i++)
     {
